Revert "Merge pull request #1682 from BrainFPV/enh_logging"
This reverts commit 749e5330b00721d7a9012191cdff1e6d58de2245, reversing
changes made to f741c8930157548fc18831b6a8378cd2a39d2a05.

+++ flight/PiOS/Common/pios_streamfs.c
@@ -110,7 +110,7 @@ struct streamfs_state {
  * @brief Return the offset in flash of a particular slot within an arena
  * @return address of the requested slot
  */
-static uintptr_t streamfs_get_addr(const struct streamfs_state *streamfs, uint16_t arena_id, uint16_t arena_offset)
+static uintptr_t streamfs_get_addr(const struct streamfs_state *streamfs, uint8_t arena_id, uint16_t arena_offset)
 {
 	PIOS_Assert(arena_id < (streamfs->partition_size / streamfs->cfg->arena_size));
 	PIOS_Assert(arena_offset < streamfs->cfg->arena_size);
@@ -135,7 +135,7 @@ struct streamfs_footer {
  * @return 0 if success, < 0 on failure
  * @note Must be called while holding the flash transaction lock
  */
-static int32_t streamfs_erase_arena(const struct streamfs_state *streamfs, uint16_t arena_id)
+static int32_t streamfs_erase_arena(const struct streamfs_state *streamfs, uint8_t arena_id)
 {
 	uintptr_t arena_addr = streamfs_get_addr(streamfs, arena_id, 0);
